pkg/loop: fix logger
